entries-api/filesystemfileentry-manual.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Entries API: FileSystemFileEntry manual test</title>
<link rel=help href="https://wicg.github.io/entries-api/#api-entry">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>
file_entry_test('file.txt', (t, entry) => {
  // FileSystemEntry ----------

  assert_idl_attribute(entry, 'isFile', 'FileSystemEntry has isFile attribute');
  assert_equals(typeof entry.isFile, 'boolean', 'isFile is boolean');

  assert_idl_attribute(entry, 'isDirectory', 'FileSystemEntry has isDirectory attribute');
  assert_equals(typeof entry.isDirectory, 'boolean', 'isFile is boolean');

  assert_idl_attribute(entry, 'name', 'FileSystemEntry has name attribute');
  assert_equals(typeof entry.name, 'string', 'name is a string');

  assert_idl_attribute(entry, 'fullPath', 'FileSystemEntry has fullPath attribute');
  assert_equals(typeof entry.fullPath, 'string', 'fullPath is a string');

  assert_idl_attribute(entry, 'filesystem', 'FileSystemEntry has filesystem attribute');

  assert_idl_attribute(entry, 'getParent', 'FileSystemEntry has filesystem attribute');
  assert_equals(typeof entry.getParent, 'function', 'FileSystemEntry has a getParent() method');

  // FileSystemFileEntry ----------

  assert_equals(typeof entry.file, 'function', 'FileSystemFileEntry has a file() method');
  t.done();
}, 'FileSystemFileEntry - interface');

file_entry_test('file.txt', (t, entry) => {
  assert_true(entry.isFile);
  assert_false(entry.isDirectory);
  assert_equals(entry.name, 'file.txt');
  assert_equals(entry.fullPath, '/upload/file.txt');
  t.done();
}, 'FileSystemFileEntry - attributes');

file_entry_test('file.txt', (t, entry) => {
  assert_throws(TypeError(), () => entry.file(), 'file() has a required argument');
  entry.file(t.step_func(file => {

    assert_class_string(file, 'File', 'file() should yield a File');
    assert_equals(entry.name, file.name, 'entry and file names should match');
    t.done();

  }), t.unreached_func('file() should not fail'));
}, 'FileSystemFileEntry - file()');

// TODO: Manual test where file is replaced with directory before file() called
</script>
